Split out post form from CreatePost component. #3
