+++ src/components/index.js
@@ -2,5 +2,6 @@ export { default as App } from './App';
 export { default as Category } from './Category'
 export { default as CreatePost } from './CreatePost'
 export { default as Navbar } from './Navbar'
+export { default as PostForm } from './PostForm'
 export { default as SelectField } from './SelectField'
 export { default as TextInput } from './TextInput'
